+++ gallery_generator/controllers/pictures.py
@@ -1,9 +1,33 @@
 import pathlib
+import re
+from typing import Iterable
+
 from exif import Image as EImage
 from PIL import Image as PILImage
 import datetime
 
 from gallery_generator.models import Picture
+
+
+def seek_pictures(
+        root: pathlib.Path,
+        extensions: Iterable[str],
+        exclude_dirs: Iterable[str]
+) -> Iterable[pathlib.Path]:
+    """Look in subdirectories of `root` for all pictures, recognized by their extension.
+
+    Return an iterable of path to pictures, relative to `root`.
+    """
+
+    r = re.compile(r'.*\.({})'.format('|'.join(extensions)))
+
+    for dir in root.iterdir():
+        if dir.name in exclude_dirs:
+            continue
+
+        for f in dir.glob('*.*'):
+            if r.match(str(f)):
+                yield f.relative_to(root)
 
 
 def create_picture_object(root: pathlib.Path, path: pathlib.Path) -> Picture:
