+++ tests/tests_crawl.py
@@ -1,9 +1,8 @@
 from tests import GCTestCase
 from sqlalchemy import func, select
 
-from gallery_generator.controllers.files import seek_pictures
 from gallery_generator import CONFIG
-from gallery_generator.controllers.pictures import create_picture_object
+from gallery_generator.controllers.pictures import create_picture_object, seek_pictures
 from gallery_generator.models import Tag, Category, Picture, tag_picture_at
 from gallery_generator.controllers.tags import TagManager
 from gallery_generator.scripts.crawl import command_crawl
